Allow running the open loan e2e test on regtest and signet
It will be based on the `.env` file located in the root directory of
the repository.

We also temporarily set the `MIN_CONFIRMATIONS` to 2, so that the
signet version of the test can actually finish in a timely manner.
Since this constant is on the way out anyway[^1], this is okay.

[^1]: https://github.com/get10101/loan-poc/issues/55.