feat: introduce liquidation engine
the liquidation engine checks every open position on every price update (to be changed).
There are two margin calls possible and afterwards the position gets set to liquidated.
This patch only does the backend part, i.e. to get out of this situation the user (lender or borrower) will have to create a dispute, then we can go in and help with the liquidation.