feat: Include lender in multisig
We do not offer any mechanism to sign with the lender's key, but that
can be added later. With this change we ensure that there is an avenue
to spend the collateral which does involve the lender.

I originally thought I could get away with a much lighter lender
wallet, but integrating with JS/TS Bitcoin libraries was such a pain
that I gave up and reused the borrower wallet for the lender. The
lender currently only needs to be able to see the mnemonic seed phrase
and to use the Xpub.

We save the lender's Xpub in the clear to local storage because it's
just a public key. This allows us to approve loan requests without
having to enter the passphrase, which is better UX.

With this patch, I've also fixed the e2e test (at least on signet).