chore(frontend): Do not fall back on relative path for APIs
The relative path is wrong if we use something like `vite` during
local development. We should expect the environment variable to be set
in all environments.

If we fall back on the wrong URL, the frontend will, for example, not
get a price from the backend, leading to subtle bugs.