chore: ignore double dipping
In case we receive a notification about a payment which invoice has been already been fully paid we ignore it, i.e. we do not move additional funds onto the card. The reason for this is that we don't have controll over who calls this endpoint and a maliocious or faulty caller could try to move more funds onto his card for the same paid invoice