fix(mempool): Do not delay confirmation of TXs and TXOs
When we calculate confirmations for transactions or transaction
outputs that were included in a block according to mempool, we may
find that the actor's cached blockchain tip is slightly outdated.

In such a scenario, we now default to 1 confirmation, since according
to mempool the transaction or transaction output was just confirmed.