fix: use unique key for browser wallet storage
I first considered generating the unique key in the browser-wallet library but decided against it because it is a bit weird if the library knows about `email` and then derives a key from it for the storage. The caller wouldn't be able to relate the browser storage with the email then.
Now the caller is responsible for generating a unique storage key. This has the disadvantage that we have to do this in multiple places, but I believe this is the cleanest approach.