feat: Use a 2-of-3 multisig instead of a 2-of-4
The 2-of-4 multisig was introduced to avoid truly catastrophic
failure (both borrower _and_ lender losing their keys), but it's
really hard to justify and it gives too much power to the hub.

Unfortunately, we've already opened a bunch of contracts, so we need
to introduce this change in a backwards-compatible manner. Hence, we
introduce the concept of a `ContractVersion`. With the
`ContractVersion` persisted in the database we know what descriptor to
reconstruct for each contract.

For now we can simply branch inline based on the `ContractVersion`,
but if we choose to follow this pattern to introduce the DLC-based
collateral contract, it won't be so simple.