fix(mempool): Give up on using track-address API
Tracking addresses via the Mempool WebSocket is nice if you pay for
the Mempool Enterprise plan, but we don't, so it's not nice. We were
running into the fact that `track-address` only allows tracking one
address per WS connection. We learnt about the `track-addresses` API
via their support Matrix channel, but that one is limited to 10
addresses for free.

I considered spawning a WS connection per contract, but that is quite
hard to manage, and Mempool probably limits the number of connections
per IP anyway.

I finally settled on polling the REST API every time we learn that a
new block has arrived. To avoid rate limits, I've spaced out the
requests by half a second.