fix: Add functionality to avoid spamming users with emails
This was introduced primarily to prevent sending the
loan-collateralised email to the lender more than once, but we have
introduced the ability to mark any one-time email as sent.

Furthermore, we also make sure that we don't even try to send a
loan-collateralised email if `db::contracts::update_collateral` is
called and it reports that the loan didn't just become collateralized.