fix: Make sure that we don't keep tracking reused contract addresses
If we ever end up generating a contract address that has already been
used, we should not proceed. This was already the case before this
patch, sort of by chance.

But we were still tracking this reused contract address in the
`mempool::Actor`, thus generating lots of noisy errors.

We now only track a new contract address if we know that it has never
been used before.