feat(mempool): Redesign mempool Actor to call Mempool's API less
We only go against the `get_address_transactions` REST API after a hub
restart or if we get a completely new collateral address.

We rely on the WS subscription to new blocks to learn if collateral
outputs have been sufficiently confirmed. On WS reconnect we get the
last 8 blocks anyway, so we should not miss any TXs even if we're
disconnected for a long time. We could further optimise this by
tracking which blocks have already been processed by the
`mempool::Actor`.

Also, we are now correctly resubscribing to new blocks on the WS after
a reconnect!

To be even safer, we've introduced the `backon` dependency, which
allows us to retry functions with exponential backoff. With this we
can keep retrying against mempool with a smart, dynamic timeout in
between retries.

We should consider using `backon` in more places!