fix(borrower): Do not assume fixed index when claiming collateral
After fixing the derivation index to 0 for all borrower keys, we saw
at least one borrower fail to spend the collateral. It seems like this
borrower's wallet used an index greater than 0 (before the update),
and now it can never re-derive the correct matching key pair, since we
force it to use a 0 index.

To rectify this we can let the hub tell the borrower which
`borrower_pk` is expected (it's already embedded in the Miniscript
descriptor, but parsing is fiddly), so that the borrower's wallet can
gradually increase the derivation index until a key pair with a
matching public key is generated.