feat: Allow lenders manually recover collateral
The flow is extremely simple:

- We create an entry in the new `manual_collateral_recovery` table,
specifying how much the lender is supposed to get.

- The lender opens the contract-details page and has a
recover-collateral action available.

- The lender completes the flow (much like a regular liquidation).

To clean up we then have to manually delete the entry from the
`manual_collateral_recovery` table.

In some cases the lender should get 0 (if the loan has already been
repaid). This is supported too.

---

A few other things happened with this patch:

- Some frontend bugs were fixed. In particular, the borrower
collateral value after funding was incorrect, because it included the
origination fee.

- I had to refactor some things so I introduced and ran `just
check-frontend` to ensure that I didn't break anything. We should run
this on CI.