feat(lender): introduce loan reserve
The lender can define a loan offer reserve when creating an offer. In total borrowers cannot request more than this amount combined, e.g. if a loan amount ranges from 1 to 100 and the reserve is 200 then there can be a maximum of 2 loans with 100 each (or different combinations). The remaining reserve balance is calculated based on the contracts in the db and is not a field persistant in the db. We take all contracts in account which are not cancelled or expired.