feat: introduce user connection details middleware
The issue we are working around here is that if we run the hub behind nginx the user's ip is always localhost (as forwarded by nginx). In our nginx setting we can however set a header 'X-Real-IP' which holds the actual ip. However! if run locally in dev mode, this ip is not available, hence, we fallback to `ConnectInfo<SocketAddr>` which is the actual connection info.