fix: Fix auto-accept feature
We now make the lender provide their Xpub on loan offer creation. This
means the browser wallet must exist before creating a loan offer.

To maintain backwards-compatibility, we add an optional `lender_xpub`
column to the `loan_offers` database table. If the `lender_xpub` is
absent, the borrower will fail to request a contract based on it.

As such, we will have to populate the `lender_xpub` column for all
active offers (2 at the time of writing) in production.