chore: Circumvent instrument(err) limitation
If an error is returned by a function annotated with
`#[instrument(err)]`, all the fields are ignored[1].

We circumvent this by adding a DEBUG log event to provide the user ID
context. This event will be logged whether we encounter an error or
not. This is not perfect, but it's good enough as long as we don't
have too many users doing authentication-related things at the same
time.

[1]: https://github.com/tokio-rs/tracing/issues/3076.