chore(browser-wallet): Avoid overwriting wallet whenever possible
- Creating a new wallet should only ever happen on registration. But
you never know, so we just move the wallet to a different key if we do
find one.

- Creating a wallet from an existing mnemonic could happen on a
browser which already holds a wallet under the same key. Moving the
existing wallet to another key does make sense, just in case.

- Restoring a wallet from a backup should only happen if a wallet does
not already exist in local storage. But you never know, so we just
move the wallet to a different key if we do find one.

- Changing the encryption of a wallet necessarily happens on a browser
which already holds a wallet under the same key! Moving the existing
wallet to another key does make sense, just in case.

- We still do not allow wallet upgrades (due to the PAKE
authentication upgrade) to occur more than once.

We are being as cautious as possible with these changes, because
wallet data is incredibly valuable.