fix(browser-wallet): Only persist wallet after registration succeeds
Before this patch (and the previous one), an already registered user
could generate a second wallet after attempting to register a second
time. The registration would fail, but logging in would allow them to
use this second wallet to create contracts. Logging in on any other
device would pull the original encrypted wallet backed up by the hub.

By splitting wallet generation from persisting (and loading) the
wallet, we first check with the hub before committing to the newly
generated wallet.

Alternatively, we could first checked with the hub, but I think it's
better to keep communication with the server as low as possible.

Fixes https://github.com/lendasat/lendasat/issues/557.