feat: Let borrower/lender interact with API with API key
This patch does not include any frontend work since the frontend is
currently unstable.

A client can generate its own API key and send a hash of it to `POST
/api/keys`.

The client can then use the API key to call any endpoint that requires
authentication.

Borrowers and lenders are allowed to generate up to 5 separate API
keys. They are also able to retrieve information about them (id,
description), as well as delete them.

We insert API keys in the `test_data.rs` script for the default
borrower and lender. The API keys are "satoshi" and "lendatoshi"
respectively.