feat(hub): Let borrower reset legacy password
We have at least one borrower in production who requires a legacy
password reset: they don't remember their legacy password.

I don't want to support this in the frontend, because it's annoying
and extremely niche. So I propose to only support it in the backend.
The flow will be:

1. Borrower gets in touch about not remembering their legacy password.

2. We insert a password reset token for that user in the database and
manually send an email to the corresponding email address with the
token and some instructions.

3. The instructions explain how to call the API: we provide a link
like this one[1] and explain that they need to replace RESET_TOKEN
with the token we gave them and NEW_PASSWORD with the legacy password
they want to use now.

After that, they can run the PAKE upgrade protocol via the UI using
the new legacy password as their current password.

[1]: https://reqbin.com/9jinfbhc.