feat: Introduce borrower API accounts
Someone holding a valid API account creator API key is able to create
borrower API accounts using:

POST /api/create-api-account

The response returns some the new user ID (plus some metadata) and the
API key that can be used to interact with the REST API on behalf of
that user.

Notes:

- We have written an e2e test where we create a borrower API
account (and use the account to open and repay a loan) to prove that
the feature works.

- The changes should have no effect on the frontend, other than the fact
that the borrower Xpub is now a field that in the Contract object.

- To onboard a new API account creator, we can manually randomly
generate an API key and insert a SHA256 hash of the API key in the
`api_account_creator_api_keys` table.

- There are now 3 different authentication middleware functions:

1. `jwt_auth` for APIs that can only be called by
password-authenticated borrower accounts (only `/api/users/me` thus far).

2. `jwt_or_api_auth` for APIs that can be called by a
password-authenticated or an API-key-authenticated borrower (most
APIs).

3. `api_account_creator_auth` to create (and eventually manage) API
accounts.

- Finally, we ensure that emails are always trimmed of whitespace
before registering borrowers and lenders.