fix: use better fee rate estimator
We estimate fee based on an actual witness stack, i.e. we construct a sample transaction and pretend it is fully signed, then estimate the weight of the transaction. We then reduce one of the outputs until we reach the wanted FeeRate.
We chose to adjust the output by 100 sats every iteration which means, in the worst case, we will be overshooting by 99 sats.
The number of iterations depends on the chosen fee rate the transaction size. E.g. for a "normal" contract transaction with 1 input and 2 outputs (refund + originaton fee) and a fee rate of 13 sats/vbyte we will get to around 2,000 sats in fee or 20 iterations of our estimator.