feat: Allow borrowers and lenders to supply PK and path
We remove Xpubs from the public API of the hub. The hub no longer
demands Xpubs of borrowers and lenders. Instead, the hub expects a PK
and a derivation path. The PK is used directly in the collateral
script. The derivation path is stored to be able to remind the user at
signing time.

External users of the API are encouraged to use hardened derivation
paths, but we choose to use a normal derivation path in our browser
wallet for a better UX.

Additionally, we keep the Nsec in local storage to be able to unlock
the contract chat without asking for a user password.

Co-authored-by: Lord Lendalot <lord_lendalot@lendasat.com>