chore(hub): Rework error logging in contract routes
The pattern is:

- Use `tracing::instrument` to log on error and return, including all
the relevant context. This context can be expanded with tracing::Span
wherever needed.

- Do not log when mapping to axum::Response. This was actually
duplicating error logs wherever we were also using
`tracing::instrument` with the `ret` argument.

We can slowly apply this new, lighter pattern to all other routes.

The only downside is that error messages are not as pretty, because we
use the `Debug` implementation of `Error`. We could use something like
`thiserror` to implement `Display` conveniently, so that we can use
`err(Display)` in the `tracing::instrument` attribute.