fix: differentiate between claim transactions
When liquidating a contract, we only started tracking the transaction but lost the context if this transaction is a liquidation or a normal claim tx. It just "worked" because a few lines after starting tracking we inserted the claim transaction correctly. Nevertheless, to avoid race conditions, we add this claim transaction type to the tracking.