feat: Integrate with Bringin
The borrower has to:

1. Register with Bringin.

2. Paste their Bringin email address into the Lendasat `Integrations`
tab.

3. Get a Email from Bringin and allow Bringin to connect with
Lendasat.

With that last step, we get a Bringin API key for that borrower, which
can be used to request Bringin quotes when creating contract requests.

If the user is not yet registered with Bringin, they can still go to
the `Integrations` tab and paste the email adddress they want to
register with. After registering, Lendasat will also get their Bringin
API key.