feat: Introduce loan payout configuration for offers
All regular and existing offers are of type `LoanPayout::Direct`. This
means that the loan amount is disbursed _directly_ to the borrower.

A new kind of loan payout is introduced: `LoanPayout::Indirect`. This
type of payout is meant for online stores that wish to allow customers
to pay later, by putting up collateral instead. They effectively lend
out the value of the product or service being purchased, but no money
is transferred from lender to borrower directly, which is why this
loan payout type is called "indirect".

Additionally, for indirect payout loans, we transition to `PrincipalGiven` as soon
as the collateral is confirmed.