feat(nostr): Add nostr 0xchat support
According to NIP-17 (https://github.com/nostr-protocol/nips/blob/master/17.md) all receivers must p-tagged in the event note (kind 14). It's nowhere clearly mentioned if the sender also needs to be part of the p-tags, but testing with 0xchat I found it is expecting it so. (Note, I reverse engineered an event note created by 0xchat (which supports NIP-17))

Unfortunately using the sendPrivateMsg API did not allow me to add the senders pubkey to the p-tags - as it was removing it when building the event.

However, building the event directly using the event builder allowed me to explictely `allow_self_tagging`, thus the authors p-tag was not removed.

With these changes the lendasat nostr chat is now compatible with 0xchat. I've additionally added a subject, as this makes the chat room look a bit nicer on 0xchat.