feat: Introduce repayment plans
Fixes https://github.com/lendasat/lendasat/issues/719.

- All existing loan offers and contracts are "transformed" into bullet
loans. For existing, open contracts we create a single balloon
installment.

- When creating loan offers, lenders can choose between bullet
and (monthly) interest-only. The API related to loan
applications/requests does allow for choosing a different repayment
plan, but the UI has not been updated yet.

- Borrowers now make payments based on the next installment, rather
than the entire loan principal + interest. This can be limiting if a
borrower wants to pay more than one installment at once, but we can
work on that later. Borrowers report paying an installment.

- Lenders confirm installment payments. When the last pending
installment is paid and confirmed, the collateral can be claimed as
usual.

- Liquidation due to defaulting on a payment is now governed by the
state of all installments associated with a contract. For now we apply
a simple rule: missing an installment leads to full liquidation of the
collateral. This can be evolved later.

- We send emails when installments are due soon, when they have been
paid and when they have been confirmed.