feat(borrower): Add REST endpoint to track collateral transactions
The `GET /api/contracts/{id}/collateral-transactions` REST API
endpoint allows clients to track collateral transactions for relevant
contracts.

This feature is powered by a new `electrum` actor, which connects to
an Electrum server (our own when running against Mutinynet). The
feature is optional, based on whether ELECTRUM_URL is set or not.

We also implement using the new endpoint in the borrower webapp. The
main purpose is to give live feedback when the contract is in status
`Approved` and the borrower has just sent funds to the collateral
address.

The feature would also work in other scenarios, such as adding
collateral when the contract is already "open", but we prefer to leave
that out for now because the UI gets confusing with multiple sources
of truth for transactions and state. I think it would be a good idea
to rework our collateral tracking (persisting outputs instead of just
the `collateral_amount`; replacing the `mempool` actor fully; maybe
even reworking the state machine altogether) before continuing with
more features like this one.