feat: backend for allowing to top-up a moon card
We will charge a fee for toping up cards if not done through a loan (because we already have an origination fee for the loan).
For sake of simplicity, we will collect our fee with Moon, i.e. we let the user pay more to moon and then send an amount less fee to the card.
For that, we need to create an invoice with moon which includes our fee.