fix(hub): Ensure that instant loans have correct late penalty
The fix was to not default to `LatePenalty::FullLiquidation` when
inserting installments into the database.

Additionally, we simplify the branching logic in the
`post_contract_request` handler and ensure that the same logic is
applied when creating a contract from a loan application. Loan
applications currently cannot be of the `MoonCardInstant` type, but
it's safer to handle the future possibility.