fix: no more empty cards
we can create the invoice without creating a new card. Once the invoice was paid, we check if a card id was provided (i.e. it was a topup) or not, if no card was provided, we create a new card for the user and then move the funds. The result -> no more empty cards when a loan was not funded :)