chore: remove loan reserve in favor of loan_amount_max
the loan reserve was overly complicated to understand from the user perspective and to handle in the, instead, we only have a loan max amount. In the next patch we will reduce the loan max amount once a loan has been taken, and increase it again if a loan request was rejected, cancelled or expired.